ai_agent/checkpoint/pyproject.toml
[project]
name = "checkpoint"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.6",
    "langgraph>=0.5.0",
    "langgraph-checkpoint>=2.1.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
